Criada exceções para account e métodos para sacar, depositar e transferir valores para outras contas via endpoint

+++ src/main/java/com/senai/devinhouse/m1s09_spring_boot/service/AccountService.java
@@ -1,11 +1,21 @@
 package com.senai.devinhouse.m1s09_spring_boot.service;
 
-import com.senai.devinhouse.m1s09_spring_boot.model.Account;
+import com.senai.devinhouse.m1s09_spring_boot.controller.dto.AccountDTO;
+import com.senai.devinhouse.m1s09_spring_boot.controller.dto.AccountOperationDTO;
+import com.senai.devinhouse.m1s09_spring_boot.controller.dto.AccountTransferDTO;
+import com.senai.devinhouse.m1s09_spring_boot.controller.forms.NewTransferForm;
+import com.senai.devinhouse.m1s09_spring_boot.model.account.Account;
+import com.senai.devinhouse.m1s09_spring_boot.model.account.AccountOperations;
+import com.senai.devinhouse.m1s09_spring_boot.model.enums.OperationType;
 import com.senai.devinhouse.m1s09_spring_boot.repository.CrudRepository;
+import com.senai.devinhouse.m1s09_spring_boot.service.exceptions.AccountOperationNotValidException;
+import com.senai.devinhouse.m1s09_spring_boot.service.exceptions.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class AccountService implements CrudService<Account> {
@@ -16,12 +26,63 @@ public class AccountService implements CrudService<Account> {
         this.repository = repository;
     }
 
+    public AccountOperationDTO withdraw(Integer id, Double value) {
+        Account account = repository.findById(id).get();
+        boolean withdrawIsValid = account.withdraw(value);
+
+        if(withdrawIsValid){
+            AccountOperations accountOperations = new AccountOperations(value, account, OperationType.WITHDRAW);
+            account.getAccountOperationsList().add(accountOperations);
+            return new AccountOperationDTO(accountOperations);
+        }
+
+        throw new AccountOperationNotValidException("Withdraw could not be performed!");
+
+    }
+
+    public AccountOperationDTO deposit(Integer id, Double value) {
+        Account account = repository.findById(id).get();
+        boolean depositIsValid = account.deposit(value);
+
+        if(depositIsValid){
+            AccountOperations accountOperations = new AccountOperations(value, account, OperationType.DEPOSIT);
+            account.getAccountOperationsList().add(accountOperations);
+            return new AccountOperationDTO(accountOperations);
+        }
+
+        throw new AccountOperationNotValidException("Deposit could not be performed!");
+
+    }
+
+    public AccountTransferDTO transferValueToOtherAccount(NewTransferForm newTransfer) {
+
+        Account originAccount = repository.findById(newTransfer.getOriginAccountId()).get();
+        Account destinyAccount = repository.findById(newTransfer.getDestinyAccountId()).get();
+
+        if (originAccount.transferValueToOtherAccount(destinyAccount, newTransfer.getValue())) {
+            AccountOperations accountOperations = new AccountOperations(newTransfer.getValue(), originAccount, OperationType.TRANSFER);
+            originAccount.getAccountOperationsList().add(accountOperations);
+            return new AccountTransferDTO(accountOperations, destinyAccount);
+        }
+
+        throw new AccountOperationNotValidException("Transfer could not be performed!");
+
+    }
+
+    public Set<AccountOperations> getAccountOperations(Integer id){
+        Account account = repository.findById(id).get();
+        return account.getAccountOperationsList();
+    }
+
     @Override
     public Account create(Account account) {
-        if(repository.create(account)){
-            return account;
+
+        if (account == null) {
+            throw new NullPointerException("Account cannot be empty (null)");
         }
-        return null;
+
+        repository.create(account);
+        return account;
     }
 
     @Override
@@ -30,18 +91,36 @@ public class AccountService implements CrudService<Account> {
     }
 
     @Override
-    public Optional<Account> findById(Integer id) {
-        return repository.findById(id);
+    public Account findById(Integer id) {
+        Optional<Account> account = repository.findById(id);
+
+        if (account.isEmpty()) {
+            throw new EntityNotFoundException("Account with id " + id + " does not exists");
+        }
+        return account.get();
     }
 
     @Override
     public boolean update(Integer id, Account account) {
-        return repository.update(id, account);
+
+        if (account == null) {
+            throw new NullPointerException("Account cannot be empty (null)");
+        }
+
+        if (findById(id) != null) {
+            return repository.update(id, account);
+        }
+
+        return false;
     }
 
     @Override
     public boolean delete(Integer id) {
-        return repository.delete(id);
+
+        if (findById(id) != null) {
+            return repository.delete(id);
+        }
+        return false;
     }
 
 }
